+++ Source/OldCode/Archive/2005 TVA Code Library/Version/2.0/Source/Tva.Phasors/BpaPdcStream/AnalogDefinition.vb
@@ -28,7 +28,7 @@ Namespace BpaPdcStream
 
         End Sub
 
-        Public Sub New(ByVal parent As ConfigurationCell, ByVal dataFormat As DataFormat, ByVal index As Integer, ByVal label As String, ByVal scale As Integer, ByVal offset As Double)
+        Public Sub New(ByVal parent As ConfigurationCell, ByVal dataFormat As DataFormat, ByVal index As Integer, ByVal label As String, ByVal scale As Integer, ByVal offset As Single)
 
             MyBase.New(parent, dataFormat, index, label, 1, 0)
 
@@ -52,7 +52,7 @@ Namespace BpaPdcStream
             End Get
         End Property
 
-        Protected Overrides ReadOnly Property BodyLength() As Int16
+        Protected Overrides ReadOnly Property BodyLength() As UInt16
             Get
                 Return 0
             End Get
