+++ Source/OldCode/Archive/2005 TVA Code Library/Version/2.0/Source/Tva.Phasors/BpaPdcStream/AnalogValue.vb
@@ -22,7 +22,7 @@ Namespace BpaPdcStream
 
         Inherits AnalogValueBase
 
-        Public Sub New(ByVal parent As IDataCell, ByVal analogDefinition As IAnalogDefinition, ByVal value As Double)
+        Public Sub New(ByVal parent As IDataCell, ByVal analogDefinition As IAnalogDefinition, ByVal value As Single)
 
             MyBase.New(parent, analogDefinition, value)
